// apps/web/src/shared/api/graphql/__generated__/schema.graphql.ts
export type Maybe<T> = T | null;
// ...
export interface Scalars {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: any; output: any };
  DateTime: { input: any; output: any };
  JSON: { input: any; output: any };
  Time: { input: any; output: any };
  Upload: { input: any; output: any };
}
// ...
export interface ComponentEventEvent {
  __typename?: 'ComponentEventEvent';
  date?: Maybe<ComponentDateEventDate>;
  id: Scalars['ID']['output'];
  info?: Maybe<Scalars['String']['output']>;
  location?: Maybe<ComponentLocationLocation>;
  schedule?: Maybe<Array<Maybe<ComponentEventSchedule>>>;
}
// ...
export interface Event {
  __typename?: 'Event';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  eventInfo: ComponentEventEvent;
  image: UploadFileEntityResponse;
  mixes?: Maybe<MixRelationResponseCollection>;
  name: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface EventMixesArgs {
  filters?: InputMaybe<MixFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}

export interface EventEntity {
  __typename?: 'EventEntity';
  attributes?: Maybe<Event>;
  id?: Maybe<Scalars['ID']['output']>;
}
// ...
export interface MixRelationResponseCollection {
  __typename?: 'MixRelationResponseCollection';
  data: Array<MixEntity>;
}
// ...
export interface UploadFileEntityResponse {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
}